requirements.txt
# AWS SDK
boto3==1.35.0
   
# OpenAI
openai==2.14.0
   
# Environment variables
python-dotenv==1.0.1
   
# HTTP requests (for calendar APIs)
requests==2.32.0
   
# Twilio (SMS + SendGrid Email)
twilio==9.3.0
   
# Google Calendar API
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
   
# Microsoft Graph API (Outlook)
msal==1.31.0
   
# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
   
# Local Lambda testing
python-lambda-local==0.1.13